python-2.7.2-add-extension-suffix-to-python-config.patch
00130 #
Add "--extension-suffix" option to python-config and python-debug-config
(rhbz#732808)

This is adapted from 3.2's PEP-3149 support.

Fedora's debug build has some non-standard features (see also patch 112
above), though largely shared with Debian/Ubuntu and Windows

In particular, SO in the Makefile is currently always just ".so" for our
python 2 optimized builds, but for python 2 debug it should be '_d.so', to
distinguish the debug vs optimized ABI, following the pattern in the above
patch.

Not yet sent upstream